Use meson's python3 module to find python3 and fix exec bits on to-be-installed binaries (#1154)
* Use meson's python3 module to locate python binary

The previous method using `dependency()` used pkg-config to
typically locate `python-devel`, which is not really needed,
rather than the python binary itself. Solves #1151.

* Exec perms on files configured to become binaries

Setting executive permissions on the *.in files set to be
configured for installation as binaries, ensures that the final
binaries themselves are also executable. Solves #1148.

* Drop `chmod +x` commands from post-install script.

This is no longer needed after commit 1d4b4a12.