Add support for non-standard MPRIS ratingsExtension (#1186)
This will allow any MPRIS client to set ratings in Lollypop via the SetRating Method. This is also future proof in the sense that if/when the TrackList interface is implemented it will allow clients to set the ratings of any song in the tracklist. Support for this interface has already been added to https://github.com/JasonLG1979/gnome-shell-extensions-mediaplayer