Make desktop notifications at song changes optional and a bunch of other fixes (#1195)
Not everyone wants desktop notifications at song changes. This will make it a user option.

This also fixes searching for covers with no API key.
The encoding was wrong and attr field was wrong

And stops and exception when the track_id is None which leads to not being able to reset the db.